[transfer]
1. mu_default from 10 to 50, such that DDP policies can be more like
deterministic
2. See how DDP policies behave, no inner loop
3. no prompt information, i.e. source flag and weights
4. DDP trajectories are stochastic
5. fix np.finfo(float32) Bug
6. hidden dim from 80 to 50

+++ tutorials/GPS/libGPS/NNBuilder.py
@@ -272,12 +272,12 @@ class PhiLoss(caffe.Layer):
     def GaussianEvaluator(self, mean, covariance, testX):
         probability = 0.0
         probability = 1.0/np.sqrt(2*np.pi*covariance)*np.exp(-0.5*((mean-testX)**2)/covariance)
-        if probability < np.finfo(np.double).tiny:  
-            probability = np.finfo(np.double).tiny  
+        if probability < np.finfo(np.float32).tiny:  
+            probability = np.finfo(np.float32).tiny  
         return probability
 
     def GaussianEvaluator_m_batch(self, mean, covariance, testX):
         probability = np.zeros(len(mean))
         probability = 1.0/np.sqrt(2*np.pi*covariance)*np.exp(-0.5*((mean-testX)**2)/covariance)
-        probability = np.maximum(probability, np.finfo(np.double).tiny)
+        probability = np.maximum(probability, np.finfo(np.float32).tiny)
         return probability
